+++ extensions/system/src/System/Theme.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Pagekit\Theme;
+namespace Pagekit\System;
 
 use Pagekit\Application as App;
 use Pagekit\Module\ModuleInterface;
